atualizando ramo

+++ projeto/src/pacotePrincipal/Facade.java
@@ -31,11 +31,11 @@ public class Facade {
 		controllerUsuario.getInfoItem(nome, telefone, nomeItem, atributo);
 	}
 	
-	public void cadastrarEletronico(String nome, String telefone, String nomeItem, int preco, String plataforma) {
+	public void cadastrarEletronico(String nome, String telefone, String nomeItem, double preco, String plataforma) {
 		controllerUsuario.cadastrarEletronico(nome, telefone, nomeItem, preco, plataforma);
 	}
 
-	public void cadastrarJogoTabuleiro(String nome, String telefone, String nomeItem, int preco) {
+	public void cadastrarJogoTabuleiro(String nome, String telefone, String nomeItem, double preco) {
 		controllerUsuario.cadastrarJogoTabuleiro(nome, telefone, nomeItem, preco);
 	}
